+++ src/django/pfb_network_connectivity/urls.py
@@ -25,16 +25,16 @@ from pfb_analysis import views as analysis_views
 
 router = routers.DefaultRouter()
 
-router.register(r'organizations', user_views.OrganizationViewSet, base_name='organizations')
-router.register(r'users', user_views.PFBUserViewSet, base_name='users')
+router.register(r'organizations', user_views.OrganizationViewSet, basename='organizations')
+router.register(r'users', user_views.PFBUserViewSet, basename='users')
 router.register(r'analysis_batches', analysis_views.AnalysisBatchViewSet,
-                base_name='analysis_batches')
-router.register(r'analysis_jobs', analysis_views.AnalysisJobViewSet, base_name='analysis_jobs')
+                basename='analysis_batches')
+router.register(r'analysis_jobs', analysis_views.AnalysisJobViewSet, basename='analysis_jobs')
 router.register(r'local_upload_tasks', analysis_views.AnalysisLocalUploadTaskViewSet,
-                base_name='local_upload_tasks')
+                basename='local_upload_tasks')
 router.register(r'score_metadata', analysis_views.AnalysisScoreMetadataViewSet,
-                base_name='score_metadata')
-router.register(r'neighborhoods', analysis_views.NeighborhoodViewSet, base_name='neighborhoods')
+                basename='score_metadata')
+router.register(r'neighborhoods', analysis_views.NeighborhoodViewSet, basename='neighborhoods')
 
 
 urlpatterns = [